requirements.txt
streamlit
numpy
pandas
plotly
python-docx
kaleido